+++ example/lib/foo.dart
@@ -1,6 +1,7 @@
 import 'package:meta/meta.dart';
 import 'package:auto_data/auto_data.dart';
 import 'package:collection/collection.dart';
+import 'dart:convert';
 
 part 'foo.g.dart';