+++ example/lib/person.dart
@@ -1,5 +1,6 @@
 import 'package:meta/meta.dart';
 import 'package:auto_data/auto_data.dart';
+import 'dart:convert';
 
 part 'person.g.dart';